requirements.txt
# Core Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# AWS
boto3>=1.35.0
botocore>=1.35.0

# Agno Framework
agno>=1.2.0

# RAG Components
chromadb>=0.5.0
faiss-cpu>=1.9.0
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-aws>=0.1.0

# Document Processing
pypdf>=4.3.0
python-docx>=1.1.0
openpyxl>=3.1.0
pandas>=2.2.0
unstructured>=0.15.0
python-pptx>=0.6.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=24.1.0
httpx>=0.27.0
tenacity>=8.5.0
numpy>=1.26.0
tiktoken>=0.7.0

# Monitoring & Metrics
prometheus-client>=0.20.0

# JSON handling
orjson>=3.10.0
